front: fix cat mismatch warning appearing globally
Tabs component is a generic component, used by unrelated parts of
the application. It should not contain manage train schedule specific
logic, nor fetch data from the store.

Thus move the managetrainschedule specific
"category and rolling stock mismatch" warning
logic outside of tabs and into managetrainschedule,
preventing it from appearing in other components unexpectedly.
