+++ front/public/locales/en/translation.json
@@ -698,7 +698,6 @@
       "choose": "Choose a category",
       "noCategory": "No category"
     },
-    "categoryMismatch": "The combination of the rolling stock and the category is unusual",
     "chooseRollingStock": "Select a rolling stock to display informations",
     "comfort": "Comfort",
     "comfortAcceleration": "Comfort acceleration",
