+++ front/public/locales/fr/translation.json
@@ -698,7 +698,6 @@
       "choose": "Choisissez une catégorie",
       "noCategory": "Pas de catégorie"
     },
-    "categoryMismatch": "L’association du matériel roulant et de la catégorie est atypique",
     "chooseRollingStock": "Sélectionnez un matériel roulant pour en afficher les informations.",
     "comfort": "Confort",
     "comfortAcceleration": "Accélération « confort »",
